% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friday_wk.R
\name{friday_wk}
\alias{friday_wk}
\title{Get last day of the week where week ends on Friday}
\usage{
friday_wk(d = lubridate::today())
}
\arguments{
\item{d}{A date}
}
\value{
A date
}
\description{
Get last day of the week where week ends on Friday
}
\examples{
d <- lubridate::ymd('2024-02-06') # a tuesday
friday_wk(d)
}
