% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jo_colours.R
\name{jo_pal_grad}
\alias{jo_pal_grad}
\title{Create a gradient from primary (purple) to secondary (orange) colour}
\usage{
jo_pal_grad(n)
}
\arguments{
\item{n}{Number of colours to generate}
}
\value{
A function that will return a gradient
}
\description{
Create a gradient from primary (purple) to secondary (orange) colour
}
\examples{
jo_pal_grad(6)
}
