% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_jo.R
\name{theme_jo}
\alias{theme_jo}
\title{ggplot2 theme with my preferred settings}
\usage{
theme_jo(
  primary_colour = jo_pal$primary,
  base_colour = jo_pal$body,
  base_bg = "transparent",
  base_size = 12,
  line_colour = jo_pal$grid,
  ...
)
}
\arguments{
\item{primary_colour}{Colour for titles}

\item{base_colour}{Base text colour}

\item{base_bg}{Background colour}

\item{base_size}{Base font size}

\item{line_colour}{Base line colour}

\item{...}{Pass other arguments to theme}

\item{base_family}{Base font family}
}
\value{
A ggplot2 theme
}
\description{
ggplot2 theme with my preferred settings
}
\examples{
library(ggplot2)
ggplot(mapping = aes(x = letters[1:10], y = 1:10, fill = rep(c('o', 'e'), times = 5))) +
  geom_col() +
  labs(x = 'letter', y = 'number', fill = 'odd/even', title = 'dummy chart') +
  theme_jo()
}
