% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_quarto_html.R
\name{use_quarto_html}
\alias{use_quarto_html}
\title{Call and open the quarto html template}
\usage{
use_quarto_html(
  report_name = NULL,
  folder = "notebooks",
  open = interactive(),
  ...
)
}
\arguments{
\item{report_name}{short name for your report}

\item{folder}{folder to save the report to - defaults to notebooks}

\item{open}{should the file be opened after being created}

\item{...}{arguments to be passed to \link[usethis]{use_template}}
}
\description{
Call and open the quarto html template
}
\examples{
\donttest{
if(interactive()){
  use_quarto_html(name = "Very important report")
}
}

}
